Migrate the dedup flows to commands
