+++ analyses/tests/test_merge_mgys_duplicates.py
@@ -1,221 +1,315 @@
-import logging
+import csv
 
 import pytest
-from django.contrib.postgres.aggregates import ArrayAgg
 from django.core.management import call_command
-from django.db.models import Count
 
-from analyses.models import Assembly, Run, Sample
-from analyses.models import Study as MGnifyStudy
-from ena.models import Sample as ENASample
-from ena.models import Study as ENAStudy
-from workflows.prefect_utils.testing_utils import combine_caplog_records
+import ena.models
+from analyses.management.commands.merge_mgys_duplicates import Command
+from analyses.models import (
+    Analysis,
+    Assembly,
+    Publication,
+    Run,
+    Sample,
+    Study,
+    StudyPublication,
+    SuperStudy,
+    SuperStudyStudy,
+)
+from workflows.models import AssemblyAnalysisBatch
 
 
 @pytest.fixture
-def setup_duplicate_studies():
+def study_pair():
+    ena_study = ena.models.Study.objects.create(
+        accession="PRJEB10002",
+        additional_accessions=["ERP100002"],
+        title="Canonical ENA",
+    )
+    study_old = Study.objects.create(
+        ena_accessions=["ERP100002", "PRJEB10002"],
+        ena_study=ena_study,
+        title="Canonical study",
+        metadata={"keep": "me"},
+        results_dir="/work/original",
+    )
+    study_new = Study.objects.create(
+        ena_accessions=["ERP100002", "PRJEB10002"],
+        ena_study=ena_study,
+        title="",
+        metadata={"duplicate": "value"},
+        external_results_dir="/work/duplicate",
+    )
+    # Extra one -- just to make sure
+    extra_ena_study = ena.models.Study.objects.create(
+        accession="PRJEB10003",
+        additional_accessions=["ERP100003"],
+        title="A different study",
+    )
+    extra_ena_study.save()
+    extra_study = Study.objects.create(
+        ena_accessions=["ERP100003", "PRJEB10003"],
+        ena_study=ena_study,
+        title="Just a random study",
+        metadata={"duplicate": "value"},
+        external_results_dir="/work/duplicate",
+    )
+    extra_study.save()
+    study_old.save()
+    study_new.save()
+    return study_old, study_new, ena_study
 
-    ena_dup_a = ENAStudy.objects.create(
-        accession="PRJEB11111",
-        additional_accessions=["ERP11111"],
-        is_private=False,
-        is_suppressed=False,
-        webin_submitter="Webin-460",
+
+@pytest.mark.django_db
+def test_find_duplicated_studies_returns_matching_pair(study_pair):
+    study_old, study_new, _ = study_pair
+
+    groups = Command().find_duplicated_studies()
+
+    assert len(groups) == 1
+    assert [study.accession for study in groups[0]] == sorted(
+        [study_old.accession, study_new.accession]
     )
 
-    ena_dup_a_reversed = ENAStudy.objects.create(
-        accession="ERP11111",
-        additional_accessions=["PRJEB11111"],
-        is_private=False,
-        is_suppressed=False,
-        webin_submitter="Webin-460",
+
+@pytest.mark.django_db
+def test_find_duplicated_studies_returns_overlap_pair():
+    ena_study = ena.models.Study.objects.create(
+        accession="PRJEB10001",
+        additional_accessions=["ERP100001"],
+        title="Overlap ENA",
+    )
+    study_old = Study.objects.create(ena_study=ena_study, title="Old")
+    study_new = Study.objects.create(ena_study=ena_study, title="New")
+    study_old.ena_accessions = ["PRJEB10001", "ERP100001"]
+    study_new.ena_accessions = ["PRJEB10001"]
+    study_old.save()
+    study_new.save()
+
+    groups = Command().find_duplicated_studies()
+
+    assert len(groups) == 1
+    assert {study.id for study in groups[0]} == {study_old.id, study_new.id}
+
+
+@pytest.mark.django_db
+def test_find_duplicated_studies_ignores_studies_without_ena_accessions():
+    ena_study = ena.models.Study.objects.create(
+        accession="PRJEB19999",
+        additional_accessions=["ERP19999"],
+        title="No accession links",
+    )
+    Study.objects.create(
+        ena_study=ena_study, title="Missing accessions", ena_accessions=[]
+    )
+    Study.objects.create(
+        ena_study=ena_study, title="Missing accessions too", ena_accessions=[]
     )
 
-    ena_dup_b = ENAStudy.objects.create(
-        accession="PRJEB22222",
-        additional_accessions=["ERP22222"],
-        is_private=False,
-        is_suppressed=False,
-        webin_submitter="Webin-460",
+    groups = Command().find_duplicated_studies()
+
+    assert groups == []
+
+
+@pytest.mark.django_db
+def test_merge_mgys_duplicates_dry_run_leaves_data_unchanged(tmp_path, study_pair):
+    study_old, study_new, _ = study_pair
+    sample = Sample.objects.create(
+        ena_sample=ena.models.Sample.objects.create(
+            accession="SAMEA-DRY",
+            study=study_new.ena_study,
+        ),
+        ena_study=study_new.ena_study,
+        ena_accessions=["SAMEA-DRY"],
     )
-    ena_single = ENAStudy.objects.create(
-        accession="PRJEB33333",
-        additional_accessions=["ERP33333"],
-        is_private=False,
-        is_suppressed=False,
-        webin_submitter="Webin-460",
-    )
-
-    # Create two MGnify studies linked to the same ENA study
-    mg_dup_a_new = MGnifyStudy.objects.create(
-        id=10001, ena_study=ena_dup_a, title="Test Study 1 new"
-    )
-    mg_dup_a_old = MGnifyStudy.objects.create(
-        id=6001, ena_study=ena_dup_a_reversed, title="Test Study 1 old"
-    )
-    mg_dup_b_new = MGnifyStudy.objects.create(
-        id=10002, ena_study=ena_dup_b, title="Test Study 2 new"
-    )
-    mg_dup_b_old = MGnifyStudy.objects.create(
-        id=6002, ena_study=ena_dup_b, title="Test Study 2 old"
-    )
-
-    # Create a single MGnify study linked to a different ENA study
-    mg_single = MGnifyStudy.objects.create(
-        id=10003, ena_study=ena_single, title="Test Study single"
-    )
-
-    for study in MGnifyStudy.objects.all():
-        study.inherit_accessions_from_related_ena_object("ena_study")
-
-    # ena sample --> ena study --> run
-    ena_sample_a = ENASample.objects.create(
-        accession="ERS1111111", study=ena_dup_a, metadata={}
-    )
-    ena_sample_b = ENASample.objects.create(
-        accession="ERS2222222", study=ena_dup_b, metadata={}
-    )
-    sample_a = Sample.objects.create(ena_sample=ena_sample_a, ena_study=ena_dup_a)
-    sample_b = Sample.objects.create(ena_sample=ena_sample_b, ena_study=ena_dup_b)
-
-    # public assembly (TPA) in different studies, usually >10,000 acc
-    assembly_public = Assembly.objects.create(
-        assembly_study=mg_single, ena_study=ena_dup_a
-    )
-    run_public = Run.objects.create(
-        ena_accessions=["ERR100000"],
-        study=mg_dup_a_new,
-        ena_study=ena_dup_a,
-        sample=sample_a,
-    )
-
-    # private assembly with run and assembly in same study, usually >10,000 acc
-    assembly_private = Assembly.objects.create(
-        assembly_study=mg_dup_b_new, ena_study=ena_dup_b
-    )
-    run_private = Run.objects.create(
-        ena_accessions=["ERR200000"],
-        study=mg_dup_b_new,
-        ena_study=ena_dup_b,
-        sample=sample_b,
-    )
-
-    return {
-        "ena_dup_a": ena_dup_a,
-        "ena_dup_b": ena_dup_b,
-        "ena_single": ena_single,
-        "mg_dup_a_new": mg_dup_a_new,
-        "mg_dup_a_old": mg_dup_a_old,
-        "mg_dup_b_new": mg_dup_b_new,
-        "mg_dup_b_old": mg_dup_b_old,
-        "mg_single": mg_single,
-        "assembly_public": assembly_public,
-        "run_public": run_public,
-        "assembly_private": assembly_private,
-        "run_private": run_private,
-        "sample_a": sample_a,
-        "sample_b": sample_b,
-    }
-
-
-@pytest.mark.skip(reason="flakiness needs debugging")
-@pytest.mark.flaky(reruns=2)
-@pytest.mark.django_db(transaction=True)
-def test_deduplicate_mgys_studies(setup_duplicate_studies, caplog):
-    dup_studies = setup_duplicate_studies
-
-    duplicates = (
-        MGnifyStudy.objects.values("ena_accessions")
-        .annotate(
-            studies_with_these_accessions=Count("accession"),
-            all_mgnify_accessions=ArrayAgg("accession"),
-        )
-        .filter(studies_with_these_accessions=2)
-    )
-
-    assert len(duplicates) == 2
-
-    # Check if the ENA study accessions are in any of the ena_accessions arrays in the duplicates result
-    duplicate_accessions = [acc for dup in duplicates for acc in dup["ena_accessions"]]
-    assert dup_studies["ena_dup_a"].accession in duplicate_accessions
-    assert dup_studies["ena_dup_b"].accession in duplicate_accessions
-    assert dup_studies["ena_single"].accession not in duplicate_accessions
-
-    with caplog.at_level(logging.INFO):
-        call_command("merge_mgys_duplicates")
-        caplog_text = combine_caplog_records(caplog.records)
-
-        assert (
-            "ENA accession ['ERP11111', 'PRJEB11111'] is linked to multiple MGnify Studies"
-            in caplog_text
-        )
-        assert (
-            "ENA accession ['ERP22222', 'PRJEB22222'] is linked to multiple MGnify Studies"
-            in caplog_text
-        )
-        assert (
-            "ENA accession ['ERP33333', 'PRJEB33333'] is linked to multiple MGnify Studies:"
-            not in caplog_text
-        )
-
-
-@pytest.mark.skip(reason="flakiness needs debugging")
-@pytest.mark.flaky(reruns=5)
-@pytest.mark.django_db(transaction=True)
-def test_reassign_runs_and_assemblies(setup_duplicate_studies, caplog):
-
-    start_count = MGnifyStudy.objects.count()
-
-    with caplog.at_level(logging.INFO):
-        call_command("merge_mgys_duplicates")
-        caplog_text = combine_caplog_records(caplog.records)
-
-        # public: run and assembly in different studies
-        assert "Moving 0 assemblies from MGYS00010001 to MGYS00006001" in caplog_text
-        assert "Moving 1 runs from MGYS00010001 to MGYS00006001" in caplog_text
-        # private: run and assembly in same study
-        assert "Moving 1 assemblies from MGYS00010002 to MGYS00006002" in caplog_text
-        assert "Moving 1 runs from MGYS00010002 to MGYS00006002" in caplog_text
-
-        # Check that 2 new duplicate studies are deleted
-        assert MGnifyStudy.objects.count() == start_count - 2
-
-
-@pytest.mark.skip(reason="flakiness needs debugging")
-@pytest.mark.flaky(reruns=5)  # sometimes fails due to logging missing or something
-@pytest.mark.django_db(transaction=True)
-def test_clashing_runs_gives_warning(setup_duplicate_studies, caplog):
-    dup_studies = setup_duplicate_studies
-
-    start_count = MGnifyStudy.objects.count()
-
-    # Create duplicate clashing runs in old and new MGnify studies
+    sample.studies.add(study_new)
     Run.objects.create(
-        ena_accessions=["ERR300000"],
-        study=dup_studies["mg_dup_a_new"],
-        ena_study=dup_studies["ena_dup_a"],
-        sample=dup_studies["sample_a"],
+        ena_accessions=["ERR-DRY"],
+        study=study_new,
+        ena_study=study_new.ena_study,
+        sample=sample,
     )
-    Run.objects.create(
-        ena_accessions=["ERR300000"],
-        study=dup_studies["mg_dup_a_old"],
-        ena_study=dup_studies["ena_dup_a"],
-        sample=dup_studies["sample_a"],
+    report_path = tmp_path / "dry-run-report.csv"
+
+    call_command(
+        "merge_mgys_duplicates",
+        "--dry-run",
+        "--output-csv",
+        str(report_path),
     )
-    with caplog.at_level(logging.INFO):
-        call_command("merge_mgys_duplicates")
-        caplog_text = combine_caplog_records(caplog.records)
 
-        assert (
-            "DUPLICATE RUNS FOUND IN BOTH STUDIES: old MGYS00006001 and new MGYS00010001"
-            in caplog_text
-        )
+    assert Study.objects.filter(id=study_old.id).exists()
+    assert Study.objects.filter(id=study_new.id).exists()
+    assert (
+        Run.objects.get(ena_accessions__contains=["ERR-DRY"]).study_id == study_new.id
+    )
+    with report_path.open(newline="", encoding="utf-8") as handle:
+        rows = list(csv.DictReader(handle))
+    assert len(rows) == 1
+    assert rows[0]["status"] == "dry_run"
+    assert rows[0]["canonical_study"] == study_old.accession
 
-        assert "Deleted 1 duplicate runs from old study MGYS00006001" in caplog_text
 
-        assert (
-            "Deleting ENA study PRJEB11111 as it is no longer linked to any MGnify studies."
-            in caplog_text
-        )
-        # this time MGYS00010001 is skipped and not deleted
-        assert MGnifyStudy.objects.count() == start_count - 2
+@pytest.mark.django_db
+def test_merge_mgys_duplicates_rewires_relations(tmp_path, study_pair):
+    study_old, study_new, ena_study = study_pair
+    study_old.title = ""
+    study_old.metadata = {"keep": "me"}
+    study_old.results_dir = None
+    study_old.external_results_dir = None
+    study_old.features = {"has_prev6_analyses": False, "has_v6_analyses": False}
+    study_old.save()
+
+    study_new.title = "Preferred title"
+    study_new.metadata = {"duplicate": "value"}
+    study_new.results_dir = "/work/results"
+    study_new.external_results_dir = "/work/external"
+    study_new.features = {"has_prev6_analyses": True, "has_v6_analyses": False}
+    study_new.save()
+
+    ena_sample_old = ena.models.Sample.objects.create(
+        accession="SAMEA1", study=ena_study
+    )
+    mg_sample = Sample.objects.create(
+        ena_sample=ena_sample_old,
+        ena_study=ena_study,
+        ena_accessions=["SAMEA1"],
+    )
+    mg_sample.studies.add(study_new)
+
+    run = Run.objects.create(
+        ena_accessions=["ERR1"],
+        study=study_new,
+        ena_study=ena_study,
+        sample=mg_sample,
+    )
+    assembly = Assembly.objects.create(
+        ena_accessions=["ERZ1"],
+        ena_study=ena_study,
+        sample=mg_sample,
+        reads_study=study_new,
+        assembly_study=study_new,
+    )
+    analysis = Analysis.objects.create(
+        study=study_new,
+        sample=mg_sample,
+        run=run,
+        assembly=assembly,
+        ena_study=ena_study,
+    )
+    AssemblyAnalysisBatch.objects.create(
+        study=study_new,
+        batch_type="assembly_analysis",
+        workspace_dir="/tmp/study-batch",
+        total_analyses=1,
+    )
+    publication = Publication.objects.create(
+        pubmed_id=123456,
+        title="Study publication",
+        metadata={},
+    )
+    StudyPublication.objects.create(study=study_new, publication=publication)
+    super_study = SuperStudy.objects.create(
+        slug="merged-super-study", title="Super Study"
+    )
+    SuperStudyStudy.objects.create(study=study_new, super_study=super_study)
+    report_path = tmp_path / "applied-report.csv"
+
+    call_command("merge_mgys_duplicates", "--output-csv", str(report_path))
+
+    study_old.refresh_from_db()
+    mg_sample.refresh_from_db()
+    run.refresh_from_db()
+    assembly.refresh_from_db()
+    analysis.refresh_from_db()
+
+    assert not Study.objects.filter(id=study_new.id).exists()
+    assert Study.objects.count() == 2  # there is an extra one - not related to the dups
+    assert mg_sample.studies.filter(id=study_old.id).exists()
+    assert run.study_id == study_old.id
+    assert assembly.reads_study_id == study_old.id
+    assert assembly.assembly_study_id == study_old.id
+    assert analysis.study_id == study_old.accession
+    assert study_old.title == "Preferred title"
+    assert study_old.metadata == {"duplicate": "value", "keep": "me"}
+    assert study_old.results_dir == "/work/results"
+    assert study_old.external_results_dir == "/work/external"
+    assert study_old.features.has_prev6_analyses is True
+    assert StudyPublication.objects.filter(
+        study=study_old, publication=publication
+    ).exists()
+    assert SuperStudyStudy.objects.filter(
+        study=study_old, super_study=super_study
+    ).exists()
+    with report_path.open(newline="", encoding="utf-8") as handle:
+        rows = list(csv.DictReader(handle))
+    assert len(rows) == 1
+    assert rows[0]["status"] == "applied"
+    assert rows[0]["deleted_studies"] == study_new.accession
+
+
+@pytest.mark.django_db
+def test_merge_mgys_duplicates_accession_filter_limits_groups(tmp_path, study_pair):
+    target_old, target_new, _ = study_pair
+    other_ena_study = ena.models.Study.objects.create(
+        accession="PRJEB10004",
+        additional_accessions=["ERP100004"],
+    )
+    other_old = Study.objects.create(ena_study=other_ena_study, title="Other old")
+    other_new = Study.objects.create(ena_study=other_ena_study, title="Other new")
+    other_old.ena_accessions = ["PRJEB10004", "ERP100004"]
+    other_new.ena_accessions = ["PRJEB10004", "ERP100004"]
+    other_old.save()
+    other_new.save()
+
+    call_command(
+        "merge_mgys_duplicates",
+        "--accession",
+        target_old.first_accession,
+        "--output-csv",
+        str(tmp_path / "filtered-report.csv"),
+    )
+
+    assert Study.objects.filter(id=target_new.id).exists() is False
+    assert Study.objects.filter(id=other_new.id).exists() is True
+
+
+@pytest.mark.django_db
+def test_merge_mgys_duplicates_reconciles_overlapping_runs(tmp_path, study_pair):
+    study_old, study_new, ena_study = study_pair
+    ena_sample = ena.models.Sample.objects.create(
+        accession="SAMEA-RUNS", study=ena_study
+    )
+    mg_sample = Sample.objects.create(
+        ena_sample=ena_sample,
+        ena_study=ena_study,
+        ena_accessions=["SAMEA-RUNS"],
+    )
+    mg_sample.studies.add(study_old, study_new)
+
+    run_old = Run.objects.create(
+        ena_accessions=["ERR-DUPLICATE"],
+        study=study_old,
+        ena_study=ena_study,
+        sample=mg_sample,
+    )
+    run_new = Run.objects.create(
+        ena_accessions=["ERR-DUPLICATE"],
+        study=study_new,
+        ena_study=ena_study,
+        sample=mg_sample,
+    )
+
+    call_command(
+        "merge_mgys_duplicates",
+        "--output-csv",
+        str(tmp_path / "overlapping-runs-report.csv"),
+    )
+
+    remaining_runs = list(
+        Run.objects.filter(ena_accessions__contains=["ERR-DUPLICATE"])
+    )
+    assert len(remaining_runs) == 1
+    assert remaining_runs[0].id == run_new.id
+    assert remaining_runs[0].study_id == study_old.id
+    assert Run.objects.filter(id=run_old.id).exists() is False
